backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-generativeai==0.8.0
pdfplumber==0.11.0
scipy==1.14.0
numpy==2.1.0
numpy-financial==1.0.0
python-multipart==0.0.9
pydantic==2.9.0
